Amend Column Type of stages.owner_id

+++ db/migrate/20160911010848_create_stages.rb
@@ -2,7 +2,7 @@ class CreateStages < ActiveRecord::Migration[5.1]
   def change
     create_table :stages do |t|
       t.string :name, :null => true
-      t.integer :owner_id, :null => false
+      t.bigint :owner_id, :null => false
       t.timestamp :deleted_at, :index => true
       t.string :slug, :index => true, :null => false
